reshuffle isclose.py into two modules
